Cleaned up ledger item

+++ app/models/ledger_item.rb
@@ -19,19 +19,18 @@
 
 class LedgerItem < ActiveRecord::Base
   belongs_to :account
-  belongs_to :sender, :class_name => "Contact"
-  belongs_to :recipient, :class_name => "Contact"
   belongs_to :match
+  belongs_to :sender,     :class_name => 'Contact'
+  belongs_to :recipient,  :class_name => 'Contact'
   
   validates_associated      :sender, :recipient, :account
   validates_presence_of     :account, :currency, :transacted_on
   validates_numericality_of :total_amount
   validates_numericality_of :tax_amount, :allow_nil => true
-  
-  validate :must_have_valid_currency_code,
-           :tax_may_not_exceed_total,
-           :tax_may_not_have_inverse_sign_of_total,
-           :must_have_perspective_of_self
+  validates_exclusion_of    :total_amount, :in => [0]
+  validate                  :currency_must_validate
+  validate                  :tax_amount_must_validate
+  validate                  :perspective_must_validate
   
   before_create :set_transacted_on_to_today
   after_create  :create_matches
@@ -42,49 +41,50 @@ class LedgerItem < ActiveRecord::Base
   named_scope :debit,     :conditions => 'total_amount > 0'
   named_scope :credit,    :conditions => 'total_amount < 0'
   
-  # These are the short-hand named scopes used the search form
+  # These named scopes are used by the search form on the ledger items list
   named_scope :account, proc { |account|
     unless account.blank?
      { :conditions => { :account_id => account.to_i } }
     end
   }
+  
   named_scope :contact, proc { |contact|
     unless contact.blank?
-      if contact == "0"
+      if contact == '0'
         {
-          :joins => "INNER JOIN contacts AS senders ON senders.id = sender_id INNER JOIN contacts AS recipients ON recipients.id = recipient_id",
-          :conditions => ["senders.self = ? OR recipients.self = ?", true, true]
+          :joins => 'INNER JOIN contacts AS senders ON senders.id = sender_id
+                     INNER JOIN contacts AS recipients ON recipients.id = recipient_id',
+          :conditions => ['senders.self = ? OR recipients.self = ?', true, true]
         }
       else
-        { :conditions => ["sender_id = ? OR recipient_id = ?", contact, contact] }
+        { :conditions => ['sender_id = ? OR recipient_id = ?', contact, contact] }
       end
     end
   }
   named_scope :query, proc { |query| 
     unless query.blank?
       q = query.dup
-      sql, vars = "1 = 1", []
-      if query =~ /(-?[0-9]+\.[0-9]{2})/ || q =~ /(-?[0-9]+)(\s|$)/
-        sql << " AND total_amount = ?"
+      sql, vars = '1 = 1', []
+      if query =~ /(-?[0-9]+\.[0-9]{2})/
+        sql << ' AND total_amount = ?'
         vars << $1.to_f
-        q.gsub!(/ ?#{$1} ?/, "")
+        q.gsub!(/ ?#{$1} ?/, '')
       end
       unless q.empty?
-        sql << " AND (description LIKE ? OR identifier LIKE ?)"
-        vars << "%#{q}%"
-        vars << "%#{q}%"
+        sql << ' AND (description LIKE ? OR identifier LIKE ?)'
+        vars += 2.times.collect { "%#{q}%" }
       end
       { :conditions => [sql] + vars }
     end
   }
   named_scope :from_date, proc { |from|
     unless from.blank? || from[:year].blank?
-      { :conditions => ["transacted_on >= ?", Date.new(from[:year].to_i, from[:month].to_i, from[:day].to_i)] }
+      { :conditions => ['transacted_on >= ?', Date.new(from[:year].to_i, from[:month].to_i, from[:day].to_i)] }
     end
   }
   named_scope :to_date, proc { |to|
     unless to.blank? || to[:year].blank?
-      { :conditions => ["transacted_on <= ?", Date.new(to[:year].to_i, to[:month].to_i, to[:day].to_i)] }
+      { :conditions => ['transacted_on <= ?', Date.new(to[:year].to_i, to[:month].to_i, to[:day].to_i)] }
    end
   }
   
@@ -108,8 +108,14 @@ class LedgerItem < ActiveRecord::Base
     self.match.ledger_items.reject { |i| i.id == self.id }
   end
   
-  # Curreny stuff probably be moved out of here.
-  CURRENCY_SYMBOLS = { "USD" => "$", "EUR" => "€", "GBP" => "£", "CAD" => "CAD$", "JPY" => "¥"}
+  # This curreny stuff should probably be moved out of here.
+  CURRENCY_SYMBOLS = {
+    'USD' => '$',
+    'EUR' => '€',
+    'GBP' => '£',
+    'CAD' => 'C$',
+    'JPY' => '¥'
+  }
   
   def currency_symbol
     CURRENCY_SYMBOLS[self.currency]
@@ -121,38 +127,47 @@ class LedgerItem < ActiveRecord::Base
     self.transacted_on = Date.today if self.transacted_on.nil?
   end
   
-  def must_have_valid_currency_code
+  def currency_must_validate
     unless ISO4217::CODE.has_key?(self.currency)
       errors.add(:currency, 'is invalid')
     end
   end
   
-  def tax_may_not_exceed_total
-    if self.tax_amount.abs > self.total_amount.abs
+  def tax_amount_must_validate
+    tax_amount_must_not_exceed_total_amount
+    tax_amount_must_have_same_sign_as_total_amount
+  end
+  
+  def tax_amount_must_not_exceed_total_amount
+    if self.tax_amount.abs - self.total_amount.abs > 0
       errors.add(:tax_amount, 'may not exceed total amount')
     end
   end
   
-  def tax_may_not_have_inverse_sign_of_total
+  def tax_amount_must_have_same_sign_as_total_amount
     if self.tax_amount * self.total_amount < 0
-      errors.add(:tax_amount, 'may not have inverse sign of total amount')
+      errors.add(:tax_amount, 'must have same sign as total amount')
     end
   end
   
-  # Some more business logic to validate. If self is part of the transaction,
-  # i.e. is sender and/or recipient, she should not be solely on the receiving
-  # end if transacted_amount is negative and similarly, should not be solely
-  # on the sending end if the transacted amount is positive. Either case would
-  # imply we are recording in the ledger of the other party, not that of the 
-  # self!
-  #
-  # This should help ease erroneous manual entries.
-  def must_have_perspective_of_self
-    if self.debit? && self.sender && self.sender.self? && self.recipient && !self.recipient.self?
+  # If self is part of the transaction, i.e. is sender and/or recipient,
+  # she should not be solely on the receiving end if transacted_amount is 
+  # negative and similarly, should not be solely on the sending end if the
+  # transacted amount is positive. Either case would imply we are recording 
+  # in the ledger of the other party, not that of the self.
+  def perspective_must_validate
+    self_must_not_send_debit
+    self_must_not_receive_credit
+  end
+    
+  def self_must_not_receive_credit
+    if self.credit? && self.recipient && self.recipient.self? && self.sender && !self.sender.self?
       errors.add_to_base('Not set up from perspective of self')
     end
-    
-    if self.credit? && self.recipient && self.recipient.self? && self.sender && !self.sender.self?
+  end
+  
+  def self_must_not_send_debit
+    if self.debit? && self.sender && self.sender.self? && self.recipient && !self.recipient.self?
       errors.add_to_base('Not set up from perspective of self')
     end
   end
@@ -166,21 +181,23 @@ class LedgerItem < ActiveRecord::Base
   end
   
   def update_matches
-    if self.matched?
-      if self.match.ledger_items.count == 2
-        i = self.matched_ledger_items.first
-        if (i.total_amount + self.total_amount).to_f.abs > 0.01 ||
-          i.sender_id != self.recipient_id ||
-          i.recipient_id != self.sender_id
+    if self.matched? && self.changed? && !self.match_id_changed?
+      matches = self.matched_ledger_items
+      if matches.count == 1
+        # Update match
+        i = matches.first
+        if (i.total_amount + self.total_amount).to_f.abs > 0.01 || i.sender_id != self.recipient_id || i.recipient_id != self.sender_id
           i.update_attributes(
             :total_amount   => self.total_amount * -1.0,
             :sender_id      => self.recipient_id,
             :recipient_id   => self.sender_id
           )
         end
-      elsif self.match.ledger_items.count > 2
-        self.match.ledger_items.each { |i| i.update_attribute(:match_id, nil)}
+      elsif matches.count > 1
+        # Destroy matches
+        self.matched_ledger_items.each { |i| i.update_attribute(:match_id, nil) }
         self.match.destroy
+        self.update_attribute(:match_id, nil)
       end
     end
   end
